Support dark mode (#4)
In preparation for dark mode support in mobile (the TOS page is used on sign up screen). I've extracted the styles and script to listen to media query changes. Might not need a class name to toggle the dark theme (could just use the dark mode query directly). But it's not clear to me yet how the integration with mobile color scheme preference would work.